add very basic API endpoint to /tracker/
This is to figure out the way Wordpress works with adding new
endpoints. It has no current functionality, but does provide
testing interface.

Important to note: after install, have to visit the "Permalinks"
page - this act flushes the rewrite rules and necessary to enable
the correct operation. Should be able to do it programically as well,
(on activate/deactivate according to web sources), but that doesn't
seem to be working yet.